split out trait
