+++ crates/domino/src/main.rs
@@ -123,6 +123,7 @@ use shadow_rs::shadow;
 shadow!(build);
 
 use sspverif::project;
+use sspverif::project::Project;
 
 mod cli;
 use crate::cli::*;
@@ -136,17 +137,17 @@ pub(crate) struct Cli {
 }
 
 fn proofsteps() -> Result<(), project::error::Error> {
-    let project_root = project::find_project_root()?;
-    let files = project::Files::load(&project_root)?;
-    let project = project::Project::load(&files)?;
+    let project_root = project::directory::find_project_root()?;
+    let files = project::DirectoryFiles::load(&project_root)?;
+    let project = project::DirectoryProject::load(&files)?;
 
     project.proofsteps()
 }
 
 fn prove(p: &Prove) -> Result<(), project::error::Error> {
-    let project_root = project::find_project_root()?;
-    let files = project::Files::load(&project_root)?;
-    let project = project::Project::load(&files)?;
+    let project_root = project::directory::find_project_root()?;
+    let files = project::DirectoryFiles::load(&project_root)?;
+    let project = project::DirectoryProject::load(&files)?;
 
     assert!(p.proofstep.is_none() || p.proof.is_some());
 
@@ -175,9 +176,9 @@ fn explain(_game_name: &str, _dst: &Option<String>) -> Result<(), project::error
 }
 
 fn latex(l: &Latex) -> Result<(), project::error::Error> {
-    let project_root = project::find_project_root()?;
-    let files = project::Files::load(&project_root)?;
-    let project = project::Project::load(&files)?;
+    let project_root = project::directory::find_project_root()?;
+    let files = project::DirectoryFiles::load(&project_root)?;
+    let project = project::DirectoryProject::load(&files)?;
 
     let smtsolver = l
         .smtsolver
@@ -189,13 +190,13 @@ fn format(f: &Format) -> Result<(), project::error::Error> {
     if let Some(input) = &f.input {
         sspverif::format::format_file(input)?;
     } else {
-        let root = crate::project::find_project_root();
+        let root = crate::project::directory::find_project_root();
         sspverif::format::format_file(&root?)?;
     }
     Ok(())
 }
 
-fn wire_check(game_name: &str, dst_idx: usize) -> Result<(), project::error::Error> {
+fn wire_check(_game_name: &str, _dst_idx: usize) -> Result<(), project::error::Error> {
     /*
     let project_root = project::find_project_root()?;
     let files = project::Files::load(&project_root)?;
