+++ src/gamehops/equivalence/verify_fn.rs
@@ -23,7 +23,7 @@ use crate::{
 use super::EquivalenceContext;
 
 fn verify_oracle<UI: TheoremUI>(
-    project: &Project,
+    project: &impl Project,
     ui: Arc<Mutex<&mut UI>>,
     eqctx: &EquivalenceContext,
     backend: &impl SmtSolverBackend,
@@ -145,7 +145,7 @@ fn verify_oracle<UI: TheoremUI>(
 }
 
 pub fn verify<UI: TheoremUI>(
-    project: &Project,
+    project: &impl Project,
     ui: &mut UI,
     eq: &Equivalence,
     orig_theorem: &Theorem,
@@ -211,7 +211,7 @@ pub fn verify<UI: TheoremUI>(
 }
 
 pub fn verify_parallel<UI: TheoremUI + std::marker::Send>(
-    project: &Project,
+    project: &impl Project,
     ui: &mut UI,
     eq: &Equivalence,
     orig_theorem: &Theorem,
