+++ src/project/consts.rs
@@ -1,0 +1,5 @@
+pub const PROJECT_FILE: &str = "ssp.toml";
+
+pub const PACKAGES_DIR: &str = "packages";
+pub const GAMES_DIR: &str = "games";
+pub const THEOREM_DIR: &str = "theorem";
